dom/base/test/test_root_iframe.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=511084
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 511084</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=511084">Mozilla Bug 511084</a>
<p id="display"><iframe></iframe></p>
<div id="content" style="display: none">
</div>
<pre id="test">
  <script type="application/javascript">
  /** Test for Bug 511084 **/
  var doc = frames[0].document;
  doc.replaceChild(doc.createElement("iframe"), doc.documentElement);
  ok(frames[0][0] instanceof Window,
     "Should have a subframe window for a root iframe");
  </script>
</pre>
</body>
</html>
